Refactored to multirelease JAR that supports Java 11. Virtual Threads are supported only if run with Java 21. MultiManagedTaskListener moved to internal package. Signed-off-by:Ondro Mihalyi <mihalyi@omnifish.ee>
