+++ src/main/java/org/glassfish/enterprise/concurrent/internal/MultiManagedTaskListener.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package org.glassfish.enterprise.concurrent.virtualthreads;
+package org.glassfish.enterprise.concurrent.internal;
 
 import jakarta.enterprise.concurrent.ManagedExecutorService;
 import jakarta.enterprise.concurrent.ManagedTaskListener;
